+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## 2.5.0.1
 
 (In development)
+
+* Convert `DetectedCharset` to an enum.